chore: extract shared WorkQueueJobRow component from queue and parts views
